Fixed the evaluation function to be faster and produce numbers (before it was always 0), took out the what openes up max moves and replaced it with if no pieces cant be taken, do random between 0 and 1 (so other states can always over take it)
Added a commented out version of the selectAndMove method that saves the evaluation until the end and keeps a stack of states, could be useful for when we need to actually use trees. It is not computionally slower, or faster.

Added a silly checkmate printout to console for when playing computer.